All changes

+++ WebContent/resources/js/categories/index.js
@@ -2,6 +2,7 @@
 $(document).ready(function() {
 	Categories.loadAll();
 	
+	/* Reload link Object */
 	$('#reload').click(function(){
 		Categories.loadAll();
 	});
@@ -22,13 +23,59 @@ var Categories = new function() {
 		$('.data').remove();
 	};
 	
-	this.destroy = function(id){
-		$.destroy({
-			url: '/marketapp-be/resources/categories/destroy/' + id,
-			success: function(response){
-				Categories.loadFromResponse(response);
-			}
+	/* Edit category function */
+	this.edit = function(id) {
+		$.read('/marketapp-be/resources/categories/' + id.substring(1, id.length), function(response){
+			var statesdemo = {
+				state0: {
+					html: 'test 1 <br/> test1 <br/>' + response.method,
+					buttons: {Cancel: false, Next: true},
+					focus: 1,
+					submit: function(v, m, f) {
+						if(!v) {
+							return true;
+						}
+						else {
+							$.prompt.goToState('state1');							
+						}
+						return false;
+					}
+				},
+				state1: {
+					html: 'test 2 <br/> test 2 <br/>' + response.message,
+					buttons: {Back: -1, Exit: 0},
+					focus: 1,
+					submit: function(v, m, f) {
+						if (v == 0) {
+							$.prompt.close();
+						}
+						else if (v == -1) {
+							$.prompt.goToState('state0');
+						}
+						return false;
+					}
+				}
+			};
+			
+			$.prompt(statesdemo);
 		});
+	};
+	
+	/* Destroy category function */
+	this.destroy = function(id) {
+		if(confirm('Deseja excluir o registro selecionado?')) {
+			$.destroy({
+				url: '/marketapp-be/resources/categories/destroy/' + id.substring(1, id.length),
+				success: function(response) {
+					/* The JSON object already comes with all categories */
+					Categories.loadFromResponse(response);
+				}
+			});
+		}
+	};
+	
+	this.loadById = function(id){
+		
 	};
 	
 	/* This function loads all categories */
@@ -53,32 +100,24 @@ var Categories = new function() {
 			$('#message-red').fadeIn();
 		}
 		
-		/* Counter to staff the zebra logic */
-		var counter = 1;
-		
 		/* Iterate over the response object */
-		$.each(response.categories, function(index, obj) {
-			var rowClass = 'data ';
-			
-			/* Zebra table color logic */
-			if(counter%2 != 0) {
-				rowClass += 'alternate-row';
-			}
-			
-			counter++;
-			
-			var inputCheckbox = '<td><input  type="checkbox"/></td>';
+		$.each(response.categories, function(index, obj) {			
 			var idTd = '<td><a class="category" id="' + obj.id + '" href="#">' + obj.id + '</a></td>';
 			var nomeTd = '<td>' + obj.name + '</td>';
 			var descricaoTd = '<td>' + obj.description + '</td>';
-			var optionsTd = '<td class="options-width">' + 
-								'<a href="#" id="" title="Edit" class="icon-1 info-tooltip"></a>' +
-								'<a href="#" title="Delete" class="icon-2 info-tooltip"></a>' +
+			var optionsTd = '<td class="options-width" align="center">' + 
+							'<a href="/marketapp-fe/admin/stock/categories/edit.jsp?id=' + obj.id + '" id="d' + obj.id + '" title="Excluir" class="info-tooltip">Editar</a> | ' +	
+							'<a href="#" id="e' + obj.id + '" onclick="Categories.destroy(this.id);" title="Editar" class="info-tooltip">Excluir</a>' +
 							'</td>';
 
-			var newTr = '<tr class="' + rowClass + '">' + inputCheckbox + idTd + nomeTd + descricaoTd + optionsTd + '</tr>';
+			var newTr = '<tr>' + idTd + nomeTd + descricaoTd + optionsTd + '</tr>';
 			
-			$('#product-table tr:last').after(newTr);
+			$('#table tr:last').after(newTr);
 		});
+		
+		/* Zebra code */
+		$('#table tr:even').addClass('data alternate-row');
+		$('#table tr:odd').addClass('data');
+		$('#table tr:first').removeClass('data');
 	};
 };
